requirements.txt
fastapi
uvicorn[standard]
pydantic==2.*
transformers
peft
accelerate
bitsandbytes
faiss-cpu
chromadb
sentence-transformers
spacy
scikit-learn
numpy
pandas
opencv-python
pydub
python-multipart
OpenCV
